refactor: move module.cts
